Simplified header logic
