+++ include/xorshift64.h
@@ -5,8 +5,14 @@
 
 class XorShift64 : public Algorithm {
 public:
-    XorShift64(uint64_t seed);
-    uint64_t peekNext();
+    XorShift64(uint64_t seed) : Algorithm(seed) {};
+    uint64_t peekNext() {
+        uint64_t x = this->state;
+        x ^= x << 13;
+        x ^= x >> 7;
+        x ^= x << 17;
+        return x;
+    };
 };
 
 #endif
